+++ CHANGELOG.md
@@ -18,6 +18,7 @@ prior to assuming the existence of said check.
 - Entities can now be created via sensuctl.
 - Added the format `wrapped-json` to sensuctl `configure`, `list` and `info`
 commands, which is compatible with `sensuctl create`.
+- Added debug event log with all event data.
 
 ### Changed
 - Add logging around the Sensu event pipeline.
@@ -67,6 +68,7 @@ the organization they reside in.
 - Removed Linux/386 & Windows/386 e2e jobs on Travis CI & AppVeyor
 - Removed check output metric extraction e2e test, in favor of more detailed
 integration coverage.
+- Removed the `leader` package
 
 ## [2.0.0-beta.1] - 2018-05-07
 ### Added